Importing Only Parts of Bootstrap
Let's do one more thing! The _bootstrap.scss file imports every part of Bootstrap. But what if you don't need every part of Bootstrap? For example, we are not using the glyphicons. Importing that CSS is wasteful!

Let's fix this! Copy the entire contents of _bootstrap.scss. Then, in our css/ directory, create a new _bootstrap.scss file and paste that here.

As soon as we do that... all the paths are broken! Yea, that makes sense: we're in a different directory. No problem, I'll copy 'bootstrap and do a 'Find and Replace' to replace this with '~bootstrap-sass/assets/stylesheets/bootstrap. Replace everything!
Yay! Happy paths!

Next, in main.scss, import this file instead of the core one. Just, ./_bootstrap.scss. If this is all we did, nothing would change: we're still importing everything.