You may not have noticed, but all these cool new tools have made debugging... well... a nightmare.
Luckily, I happen to know that this code originally lives in main.scss on line 48:

 82 lines  assets/css/main.scss
... lines 1 - 46
.leaderboard {
    background-color: #FFDF00;
    padding: 10px;
    border-radius: 5px;
}

But really, I need my browser to help me out.

JavaScript is no better. When I click a row in the table, it logs a message to the console... which is apparently coming from rep_log.js on line 200. But, that's line 200 of the final, compiled file. If you look at the source assets/js/rep_log.js, it doesn't even have a line 200!

 8 lines  assets/js/rep_log.js
const $ = require('jquery');
const RepLogApp = require('./Components/RepLogApp');
{dollar_sign}(document).ready(function() {
... lines 5 - 6
});

Guys, this sucks! The answer, is sourcemaps. This is not a new concept: it's the idea that when your JavaScript or CSS is transformed, you somehow output a map that says which source line and file each final line comes from. Browsers already know how to read sourcemaps. So as long we output them, we'll start seeing the correct file and line when debugging.

Ok, I'm sold! Let's do it!