A Bit of Refactoring
Open webpack.config.js. First, we need to do a little bit of refactoring to make our life easier. Check out the CSS and Sass loaders: we're starting to get just a little bit of duplication... which is about to get worse. Near the top of this file, create a variable: const styleLoader set to an object with loader set to style-loader and some empty options.
Thanks to this, we can replace the style-loader strings below with this variable.
We're now using the expanded format for the loader... but this is exactly the same as before.