Here's a simple challenge: when I hover over the leaderboard, I want to show a JavaScript Bootstrap popover that gives us some info about how it works.
Open up the template: app/Resources/views/lift/index.html.twig. Add a class on the h2 that we can target in JavaScript: js-custom-popover. Then, I'll add a few other things that will help the popover: data-toggle='popover', a title and data-content.

To get this working, all we need to do is write some JavaScript to find that element and activate the popover. Normally, we would put this in rep_log.js. In fact, that's the best place to put it. Well, there, or some file that it requires. The point is, now that our JavaScript has matured, 100% of your JS code should live in an external JavaScript file. You should have no JavaScript in your templates, other than the one script tag:

 53 lines  app/Resources/views/lift/index.html.twig
... lines 1 - 47
{% block javascripts %}
... lines 49 - 50
    <script src='{{ asset('build/rep_log.js') }}'></script>
{% endblock %}
But, what if you're using Webpack on a legacy app... that still has a ton of JavaScript in your templates? You should refactor this. But, since that can be a big job, let's find out how we can still write JavaScript in a template... even though we shouldn't.