Add a new script tag right in the template. Here, code like normally: add a .ready() block, then find our .js-custom-popover elements and activate popover(). I'll pass that a few options.
So simple! It should work, right? After all, rep_log.js requires jQuery, so we should be able to use it in our template.

Well, try it. Refresh the page. Uh oh... error!

$ is not defined
This is coming from our template. What's going on? We are requiring jQuery!

Why jQuery / $ is not Global Anymore

Well... we already know what's going on. When we require jquery, it detects that it is being used in a module environment. So it does not set a global jQuery or $ variable. Instead, it just returns the jQuery function, which we set to $. But that only affects this file.

And this is good! We want each module to live in isolation like this. But, if you do need to expose a jQuery variable globally, you can.

First, since we're using a Bootstrap function on this page, require('bootstrap-sass') so the popover function is available.