Add two more lines: 'window.jQuery': 'jquery' and 'window.$': 'jquery'
Cool! So now we're covered. Re-start Webpack:

./node_modules/.bin/webpack --watch
And, refresh! Woh! What! An error! Once again, it says that $ is not defined! What the heck!

This is a total gotcha. Thanks to the ProvidePlugin(), it's actually now re-writing our code, changing the window.$ in a way so that it does not actually create a global variable anymore. This is not what we intended!