The fix is easy: change window.$ to global.$
What? We know that window is the global variable in a browser environment. Well, in Webpack, you're allowed to set things on a global variable. When you do, it figures out that you're in a browser environment, and sets that on window. Basically, this does the same thing as before, but won't be re-written by the ProvidePlugin(). Just by making this change, when we refresh, it works.