Watch... if I refresh... there's a slight delay before the table loads. That's because... we designed it that way! Open assets/js/Components/RepLogApp.js. In the constructor, we call this.loadRepLogs(), which... surprise! Loads the initial rep logs... by making an AJAX call 217 lines  assets/js/Components/RepLogApp.js ... lines 1 - 9 class RepLogApp {   constructor({dollar_sign}wrapper) { ... lines 12 - 16     this.loadRepLogs(); ... lines 18 - 33   } ... lines 35 - 44   loadRepLogs() {     $.ajax({       url: Routing.generate('rep_log_list'),     }).then(data => {       for (let repLog of data.items) {         this._addRow(repLog);       }     })   } ... lines 54 - 197 } ... lines 199 - 217 This builds the table.
That's cool! But... I'm going to be unreasonable and make our life complicated. I demand that this table load instantly! To make crazy me happy, let's update our app so that when the page loads, our JavaScript already knows what the initial rep logs are... without needing that AJAX call.

Well... if we still wrote JavaScript in our template... this would be pretty easy. We could create a JavaScript variable and use Twig to print out all the rep logs into that variable. Yep, mixing Twig and JavaScript is kinda handy.

But obviously... we can't start writing Twig code right in the middle of RepLogApp.js. Sigh, nope, we need a new way to communicate from our server to JavaScript. And there is a great... and simple solution.