Refactoring initialRepLogs to an Argument
First, remove loadRepLogs(). Instead, add initialRepLogs as a second argument to the constructor.

Whoever calls me will need to pass this in. Down below, loop over these: for (let repLog of initialRepLogs), then, this._addRow(repLog).

Ok! Who creates this object? Ah yes, it's our entry file: assets/js/rep_log.js. Dang it! We can't put Twig code here either. For now, just to see if this is working, I'll paste two hard-coded logs above this. This is the exact format returned by the AJAX endpoint. Pass logs as the second argument.

In theory, this should work. Side note: you should always worry when a programmer says:

In theory, this should work!
But actually, it does this time! Sometimes we get lucky. The table starts with the two hard-coded logs. Ok, we are close!