Trying to require router.js
The curious thing about the router.js file is that it lives in a PHP package. Instead of living in node_modules, it lives at vendor/friendsofsymfony/jsrouting-bundle/Resources/public/js/router.js. This is the exact file we're currently including in base.html.twig.

Cool! So, let's just require it like normal! const Routing = require(), then... well, a very ugly path: ../../../vendor/friendsofsymfony/jsrouting-bundle/Resources/public/js/router.js.

Phew! Long, gross-looking... but it makes sense. Try it. Refresh!

Uh oh... it did not like that:

Type error: Cannot read property navigator of undefined
It's coming from router.js, line 8. Huh. If you look at this file, well, it's minified... so it's pretty confusing. This file is built using something called the Google Closure compiler... which basically means that it does not play nicely with require(). I hope this will be fixed in the library soon - there's a pull request open for it. But right now... well, there is no simple way to require() it.

 Tip
It is possible to require() the file... but it takes some work. See bit.ly/fosjsroutingbundle-require(http://bit.ly/fosjsroutingbundle-require) if you're curious.