Next, in RepLogApp.js, on top, add const _ = require('lodash')
Go make sure it finished downloading. Perfect!

Find the clearForm() method near the bottom: this is called after we submit the form. Add {dollar_sign}form.find() to fetch the reps field, then add .val(_.random(1, 10)). To make sure that box has a value on load, head up to the constructor and add this._clearForm().
Perfect! Refresh now! Yes! There's our super random number. And it resets each time we submit. Management is going to love this.

So why the heck are we really doing this? It wasn't obvious, but the compiled rep_log.js file just got a lot bigger! It's 2.7Mb... but before, it was 1.27! Woh! That's all from lodash! We're importing every function from that library... even though we only need one!

So... question: is there a way for us to import just one thing? Well, it depends on how the library you're using is built. For lodash, it is possible! Look inside node_modules/ for lodash. The library is just a bunch of tiny functions all separated into individual files. When you require lodash, it loads a file that requires all of these. But... we could also require just a single file. In this case, number.js holds what we need... plus a few other things. We could require number.js or random.js.