Require only Part of a Library.
So... question: is there a way for us to import just one thing? Well, it depends on how the library you're using is built. For lodash, it is possible! Look inside node_modules/ for lodash. The library is just a bunch of tiny functions all separated into individual files. When you require lodash, it loads a file that requires all of these. But... we could also require just a single file. In this case, number.js holds what we need... plus a few other things. We could require number.js or random.js.

Back in our code, change to const _ = require('lodash/number')

In the browser, everything still works. And the package is back down to 1.3Mb!

That's great! And we're now ready to talk about the require function's hipper, better-looking, and younger sibling: import.