When it comes to using require() and module.exports, well, we rock. So now... let's completely change things! Woohoo!
Let me tell you a story of bravery, cunning and JavaScript... simplified to fit into 20 seconds. require() and module.exports basically come from Node.js. They're not part of ECMAScript - so not officially part of the JavaScript language. But, if you're in Node, you know that these are available to use.

Hello import

Webpack supports the original require() and module.exports and the ES6 syntax. But, import and export have a few subtle advantages... other than being newer and shinier.

But mostly, they work the same! Instead of const _ = require(), say import _ from 'lodash'. When we do, Webpack is still happy... and our app still works!