Using import Everywhere!
Let's change the rest of our code to use import! Since this is a chore, I'll speed through part of this. For the CSS, we don't need the value it imports. So, just say import then the filename.

Keep going: inside rep_log.js, change require() to import. And then in login.js. And finally layout.js.
After all that... Webpack is still happy! The page still works... and our code is trendier than ever!