Extracting the Manifest
Anyways, the fix is to move that Webpack bootstrap code out of layout.js... so it can happily remain unchanged even when the internal module ids change. To do that, open webpack.config.js and go to the CommonsChunkPlugin. Change the name option to an array, with layout and a new entry called manifest.

I'll move my first comment down a bit, and then add a new comment above manifest:

Dumps the manifest in a separate file.

Let's see what this does! Restart Webpack:
./node_modules/.bin/webpack --watch

Ok, we still have a layout.js file... but now we also have a tiny manifest.js! Thanks to this change, layout.js is just our code. And the tiny manifest.js contains the Webpack bootstrap. If the ids change now, the user will only need to re-download that small file.

Of course, to get this to work... we need to add another script tag! Open base.html.twig. Add a second script tag pointing to manifest.js.
Phew! I know, that was confusing. But our setup rocks! Thanks to CommonsChunkPlugin, anything in layout.js, will not be duplicated in the other files.

Next! Let's learn about the webpack-dev-server.