Current Setup: style-loader
Now open up webpack.config.js. Remember: CSS is processed via loaders. If a file ends in .css, it goes through the css-loader and then the style-loader:
125 lines  webpack.config.js
... lines 1 - 7
const styleLoader = {
    loader: 'style-loader',
    options: {
        sourceMap: true
    }
};
const cssLoader = {
    loader: 'css-loader',
    options: {
        sourceMap: true
    }
};
... lines 20 - 32
module.exports = {
... lines 34 - 43
  module: {
    rules: [
... lines 46 - 55
      {
        test: /\.css$/,
        use: [
          styleLoader,
          cssLoader,
      ]
      },
... lines 63 - 93
    ]
  },
..lines 96 - 123
};
For Sass, it's basically the same: the sass-loader, resolve-url-loader, then the same css-loader and style-loader:
125 lines  webpack.config.js
... lines 1 - 7
const styleLoader = {
    loader: 'style-loader',
    options: {
        sourceMap: true
    }
};
const cssLoader = {
    loader: 'css-loader',
    options: {
        sourceMap: true
    }
};
const sassLoader = {
    loader: 'sass-loader',
    options: {
        sourceMap: true
    }
};
const resolveUrlLoader = {
    loader: 'resolve-url-loader',
    options: {
        sourceMap: true
    }
};
module.exports = {
... lines 34 - 43
  module: {
    rules: [
... lines 46 - 62
      {
        test: /\.scss$/,
        use: [
          styleLoader,
          cssLoader,
          resolveUrlLoader,
          sassLoader,
        ]
      },
... lines 72 - 93
    ]
  },
... lines 96 - 123
};
The style-loader is the key: it embeds the CSS so that it's added to the page as a style tag. Basically, we need to replace the style-loader with something that, instead, outputs a real CSS file. That's exactly what the extract-text-webpack-plugin does!